1. Renamed sub_address_outputs -> sub_outputs_payload 2. Added a latest output id for testing
