+++ tests/client/api_events/test_outputs_payloads.c
@@ -4,14 +4,14 @@
 #include <inttypes.h>
 #include <time.h>
 
-#include "client/api/events/sub_address_outputs.h"
+#include "client/api/events/sub_outputs_payload.h"
 #include "events_test_config.h"
 #include "unity/unity.h"
 
 bool test_completed = false;
 char const *const test_bech32 = "atoi1qqs7y6ec5vcg6cnz46vjrar2epc52lhksyar3a4zua7fg7ca08y5ymep8aa";
 char const *const test_ed25519 = "21e26b38a3308d6262ae9921f46ac871457ef6813a38f6a2e77c947b1d79c942";
-char const *const test_output_id = "0c0ef1ad9cc0e6d881fc5e0c27eb2ec009ef306a45b80d0d69dfe3c58ac95b930000";
+char const *const test_output_id = "3912942d1cb588d8091eff2069bdd797a0a834739dc8ea550e35fb0dc8609c820000";
 
 void setUp(void) {}
